requirements.txt
streamlit
fastapi
uvicorn
requests
pydantic-settings
python-dotenv
firecrawl-py
google-genai